Allow deploy gae to copy yaml file from a directory or form a direct file path.
