+++ deploy-gae-build-task/deploy-gae.ts
@@ -12,17 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/**
- * @fileoverview This is the main script run by the deploy-gae-build-task.
- *   It takes input from the TFS build task GUI (defined by task.json).
- *   It then, if asked, copies the YAML file from the source folder to the
- *   deployment path. I writes a credential file from the endpoint
- *   authorization, and calls gcloud beta app deploy with various parameters.
- * @author przybjw@google.com (Jim Przybylinski)
- */
 import * as exec from 'common/exec-options';
 import {isoNowString} from 'common/format';
-import {catchAll} from 'common/handle-rejection';
 import * as path from 'path';
 import * as task from 'vsts-task-lib/task';
 import {IExecOptions, IExecResult, ToolRunner} from 'vsts-task-lib/toolrunner';
@@ -33,61 +24,65 @@ import Endpoint = exec.Endpoint;
 import TaskResult = task.TaskResult;
 
 /**
- * Runs the script
+ * @fileoverview This is the main logic of the Deploy to GAE task. It copies
+ * the YAML file from the source folder to the deployment path, writes a
+ * credential file from the endpoint authorization, and calls gcloud beta app
+ * deploy with various parameters.
+ * @author przybjw@google.com (Jim Przybylinski)
  */
-async function run(): Promise<void> {
+
+export interface RunOptions {
+  deploymentPath: string;
+  yamlFileName: string;
+  versionInput?: string;
+  endpoint: Endpoint;
+  storageBucket?: string;
+  copyYaml: boolean;
+  yamlSource?: string;
+  promote: boolean;
+  stopPrevious?: boolean;
+}
+
+export async function deployGae(runOptions: RunOptions): Promise<void> {
+
   // Check that gcloud exists.
   const gcloudPath = task.which('gcloud', true);
   checkGcloudVersion(gcloudPath);
 
-  // Get inputs from GUI.
-  // The id of the GCP service endpoint to get the credentials from.
-  const endpointId = task.getInput('serviceEndpoint', true);
-  // The path of the deployment files.
-  const deploymentPath = task.getPathInput('deploymentPath', true);
-  // The name of the YAML file we want to run on.
-  const yamlFileName = task.getInput('yamlFileName', true);
-  // If true, copy the YAML file from the source folder to the deployment
-  // path.
-  const copyYaml = task.getBoolInput('copyYaml', true);
-  // The source folder the YAML file is in.
-  const sourceFolder = task.getPathInput('sourceFolder', copyYaml);
-  // The storage bucket to send to the --bucket parameter.
-  const storageBucket = task.getInput('storageBucket', false);
-  // The version to deploy.
-  const versionInput = task.getInput('version', false);
-  // Toggle between --promote and --no-promote parameters.
-  const promote = task.getBoolInput('promote', true);
-  // Toggle between --stop-previous-version and --no-stop-previous-version.
-  const stopPrevious = task.getBoolInput('stopPrevious', promote);
+  const deploymentPath = runOptions.deploymentPath;
+  const yamlFileName = runOptions.yamlFileName;
+  const storageBucket = runOptions.storageBucket;
 
   // Move YAML.
   const yamlPath = path.join(deploymentPath, yamlFileName);
-  if (copyYaml) {
-    task.cp(path.join(sourceFolder, yamlFileName), deploymentPath);
+  if (runOptions.copyYaml) {
+    const yamlSource = runOptions.yamlSource;
+    const appendedSource = path.join(yamlSource, yamlFileName);
+    if (yamlSource.endsWith(yamlFileName) && !task.exist(appendedSource)) {
+      task.cp(yamlSource, deploymentPath);
+    } else {
+      task.cp(appendedSource, deploymentPath);
+    }
   }
 
   // Set version
-  const version = versionInput || isoNowString();
+  const version = runOptions.versionInput || isoNowString();
 
-  // Set up the key file from the certificate parameter of the service
-  // endpoint.
-  const endpoint = new Endpoint(endpointId);
-
+  const endpoint = runOptions.endpoint;
   // Set gcloud arguments.
   const projectArg = endpoint.projectParam;
   const credentialArg = Endpoint.credentialParam;
 
   const gcloud: ToolRunner =
-      task.tool(gcloudPath)
-          .line('beta app deploy --quiet --verbosity=info')
-          .arg([ `"${yamlPath}"`, credentialArg, projectArg ])
-          .argIf(version, `--version="${version}"`)
-          .argIf(storageBucket, `--bucket="${storageBucket}"`)
-          .argIf(promote, '--promote')
-          .argIf(!promote, '--no-promote')
-          .argIf(promote && stopPrevious, '--stop-previous-version')
-          .argIf(promote && !stopPrevious, '--no-stop-previous-version');
+    task.tool(gcloudPath)
+      .line('beta app deploy --quiet --verbosity=info')
+      .arg([`"${yamlPath}"`, credentialArg, projectArg])
+      .argIf(version, `--version="${version}"`)
+      .argIf(storageBucket, `--bucket="${storageBucket}"`)
+      .argIf(runOptions.promote, '--promote')
+      .argIf(!runOptions.promote, '--no-promote')
+      .argIf(runOptions.promote && runOptions.stopPrevious, '--stop-previous-version')
+      .argIf(runOptions.promote && !runOptions.stopPrevious, '--no-stop-previous-version');
 
   const execOptions: IExecOptions = exec.getDefaultExecOptions();
 
@@ -98,6 +93,7 @@ async function run(): Promise<void> {
     task.setResult(TaskResult.Succeeded, 'Deployment succeeded');
   });
 }
+
 
 function checkGcloudVersion(gcloudPath: string): void {
   interface GcloudVersion {
@@ -118,5 +114,3 @@ function checkGcloudVersion(gcloudPath: string): void {
     throw new Error(strings.noGcloudBetaError);
   }
 }
-
-catchAll(run());
