+++ deploy-gae-build-task/legacyTests/deploy-app.ts
@@ -1,4 +1,4 @@
-﻿// Copyright 2017 Google Inc. All Rights Reserved.
+// Copyright 2017 Google Inc. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0
 // you may not use this file except in compliance with the License.
@@ -14,26 +14,46 @@
 
 /**
  * @fileoverview This is a test script for the deploy-gae-build-task with
- *   an old version of gcloud.
+ *   normal inputs.
  * @author przybjw@google.com (Jim Przybylinski)
  */
 
+
 import * as mock from 'common/register-mocks';
 import * as path from 'path';
-import { TaskLibAnswers, TaskLibAnswerExecResult} from 'vsts-task-lib/mock-answer';
+import {TaskLibAnswers} from 'vsts-task-lib/mock-answer';
 import {TaskMockRunner} from 'vsts-task-lib/mock-run';
 
-const taskPath = path.join(__dirname, '..', 'deploy-gae.js');
-const runner = new TaskMockRunner(taskPath);
-const answers: TaskLibAnswers = mock.getDefaultAnswers();
-answers.exec[mock.gcloudVersionExecString] = {
-  code: 0,
-  stdout: '{\n' +
-      '    "Google Cloud SDK": "145.0.0",\n' +
-      '    "beta": "2016.01.12"\n' +
-      '}\n',
-} as TaskLibAnswerExecResult;
+import * as strings from './test-strings';
 
+const taskPath = path.join(__dirname, '..', 'run.js');
+const runner = new TaskMockRunner(taskPath);
+
+const deployPath = path.resolve('Test', 'deploy');
+
+runner.setInput('serviceEndpoint', 'endpoint');
+runner.setInput('deploymentPath', deployPath);
+runner.setInput('yamlFileName', 'app.yaml');
+runner.setInput('copyYaml', 'false');
+runner.setInput('promote', 'true');
+runner.setInput('stopPrevious', 'true');
+
+const execString = [
+  mock.gcloudPath,
+  'beta app deploy --quiet --verbosity=info',
+  strings.yamlParam,
+  strings.credentialParam,
+  strings.projectParam,
+  strings.versionParam,
+  '--promote',
+  '--stop-previous-version',
+].join(' ');
+
+const answers: TaskLibAnswers = mock.getDefaultAnswers();
+answers.exec[execString] = {
+  'code': 0,
+  'stdout': '[gcloud output]'
+};
 runner.setAnswers(answers);
 mock.registerCommonMocks(runner);
 runner.run();
