+++ deploy-gae-build-task/legacyTests/gcloud-fails.ts
@@ -25,7 +25,7 @@ import {TaskMockRunner} from 'vsts-task-lib/mock-run';
 
 import * as strings from './test-strings';
 
-const taskPath = path.join(__dirname, '..', 'deploy-gae.js');
+const taskPath = path.join(__dirname, '..', 'run.js');
 const runner = new TaskMockRunner(taskPath);
 
 runner.setInput('serviceEndpoint', 'endpoint');
