+++ deploy-gae-build-task/legacyTests/missing-copy-yaml.ts
@@ -14,50 +14,28 @@
 
 /**
  * @fileoverview This is a test script for the deploy-gae-build-task with
- *   normal inputs and setting a storage bucket.
+ *   normal inputs.
  * @author przybjw@google.com (Jim Przybylinski)
  */
 
-import * as mock from 'common/register-mocks';
+import {getDefaultAnswers, registerCommonMocks} from 'common/register-mocks';
 import * as path from 'path';
 import {TaskLibAnswers} from 'vsts-task-lib/mock-answer';
 import {TaskMockRunner} from 'vsts-task-lib/mock-run';
 
-import * as strings from './test-strings';
-
-const taskPath = path.join(__dirname, '..', 'deploy-gae.js');
+const taskPath = path.join(__dirname, '..', 'run.js');
 const runner = new TaskMockRunner(taskPath);
 
 const deployPath = path.resolve('Test', 'deploy');
 
-
 runner.setInput('serviceEndpoint', 'endpoint');
 runner.setInput('deploymentPath', deployPath);
 runner.setInput('yamlFileName', 'app.yaml');
-runner.setInput('copyYaml', 'false');
-runner.setInput('storageBucket', 'bucketname');
 runner.setInput('promote', 'true');
 runner.setInput('stopPrevious', 'true');
 
-const execString = [
-  mock.gcloudPath,
-  'beta app deploy --quiet --verbosity=info',
-  strings.yamlParam,
-  strings.credentialParam,
-  strings.projectParam,
-  strings.versionParam,
-  '--bucket="bucketname"',
-  '--promote',
-  '--stop-previous-version',
-].join(' ');
-
-const answers: TaskLibAnswers = mock.getDefaultAnswers();
-answers.exec[execString] = {
-  'code': 0,
-  'stdout': '[gcloud output]'
-};
+const answers: TaskLibAnswers = getDefaultAnswers();
 
 runner.setAnswers(answers);
-mock.registerCommonMocks(runner);
-
+registerCommonMocks(runner);
 runner.run();
