+++ deploy-gae-build-task/legacyTests/missing-source-folder.ts
@@ -14,21 +14,28 @@
 
 /**
  * @fileoverview This is a test script for the deploy-gae-build-task with
- *   a missing gcloud.
+ *   normal inputs.
  * @author przybjw@google.com (Jim Przybylinski)
  */
 
-import {registerCommonMocks} from 'common/register-mocks';
+import {getDefaultAnswers, registerCommonMocks} from 'common/register-mocks';
 import * as path from 'path';
 import {TaskLibAnswers} from 'vsts-task-lib/mock-answer';
 import {TaskMockRunner} from 'vsts-task-lib/mock-run';
 
-const taskPath = path.join(__dirname, '..', 'deploy-gae.js');
+const taskPath = path.join(__dirname, '..', 'run.js');
 const runner = new TaskMockRunner(taskPath);
-const answers = {
-  'which': {'gcloud': 'gcloud'},
-  'checkPath': {'gcloud': false}
-} as TaskLibAnswers;
+
+const deployPath = path.resolve('Test', 'deploy');
+
+runner.setInput('serviceEndpoint', 'endpoint');
+runner.setInput('deploymentPath', deployPath);
+runner.setInput('yamlFileName', 'app.yaml');
+runner.setInput('copyYaml', 'true');
+runner.setInput('promote', 'true');
+runner.setInput('stopPrevious', 'true');
+
+const answers: TaskLibAnswers = getDefaultAnswers();
 
 runner.setAnswers(answers);
 registerCommonMocks(runner);
