+++ deploy-gae-build-task/legacyTests/missing-yaml-file-name.ts
@@ -14,7 +14,7 @@
 
 /**
  * @fileoverview This is a test script for the deploy-gae-build-task with
- *   a missing service endpoint input.
+ *   normal inputs.
  * @author przybjw@google.com (Jim Przybylinski)
  */
 
@@ -23,13 +23,13 @@ import * as path from 'path';
 import {TaskLibAnswers} from 'vsts-task-lib/mock-answer';
 import {TaskMockRunner} from 'vsts-task-lib/mock-run';
 
-const taskPath = path.join(__dirname, '..', 'deploy-gae.js');
+const taskPath = path.join(__dirname, '..', 'run.js');
 const runner = new TaskMockRunner(taskPath);
 
 const deployPath = path.resolve('Test', 'deploy');
 
+runner.setInput('serviceEndpoint', 'endpoint');
 runner.setInput('deploymentPath', deployPath);
-runner.setInput('yamlFileName', 'app.yaml');
 runner.setInput('copyYaml', 'false');
 runner.setInput('promote', 'true');
 runner.setInput('stopPrevious', 'true');
@@ -38,5 +38,4 @@ const answers: TaskLibAnswers = getDefaultAnswers();
 
 runner.setAnswers(answers);
 registerCommonMocks(runner);
-
 runner.run();
