+++ deploy-gae-build-task/legacyTests/no-gcloud.ts
@@ -14,27 +14,21 @@
 
 /**
  * @fileoverview This is a test script for the deploy-gae-build-task with
- *   normal inputs.
+ *   a missing gcloud.
  * @author przybjw@google.com (Jim Przybylinski)
  */
 
-import {getDefaultAnswers, registerCommonMocks} from 'common/register-mocks';
+import {registerCommonMocks} from 'common/register-mocks';
 import * as path from 'path';
 import {TaskLibAnswers} from 'vsts-task-lib/mock-answer';
 import {TaskMockRunner} from 'vsts-task-lib/mock-run';
 
-const taskPath = path.join(__dirname, '..', 'deploy-gae.js');
+const taskPath = path.join(__dirname, '..', 'run.js');
 const runner = new TaskMockRunner(taskPath);
-
-const deployPath = path.resolve('Test', 'deploy');
-
-runner.setInput('serviceEndpoint', 'endpoint');
-runner.setInput('deploymentPath', deployPath);
-runner.setInput('copyYaml', 'false');
-runner.setInput('promote', 'true');
-runner.setInput('stopPrevious', 'true');
-
-const answers: TaskLibAnswers = getDefaultAnswers();
+const answers = {
+  'which': {'gcloud': 'gcloud'},
+  'checkPath': {'gcloud': false}
+} as TaskLibAnswers;
 
 runner.setAnswers(answers);
 registerCommonMocks(runner);
