+++ deploy-gae-build-task/task.json
@@ -120,7 +120,7 @@
   ],
   "execution": {
     "Node": {
-      "target": "deploy-gae.js"
+      "target": "run.js"
     }
   }
 }
